Add tests, readme
